Split single module file out into directory module
